requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0
nltk>=3.8.1
pymorphy3>=2.0.0
joblib>=1.3.0
openai-whisper